Ask user for authorisation token.
Program now brings user to an index page which asks user for a Github user to be searched and an access token to allow the program to access the GitHub api. Multiple users can be explored but the token must be re-entered.
Code tidied up.
Image of final output uploaded